+++ packages/remix-testing/__tests__/stub-test.tsx
@@ -1,12 +1,20 @@
 import * as React from "react";
-import { render, screen } from "@testing-library/react";
-import { unstable_createRemixStub } from "@remix-run/testing";
-import { Outlet, useLoaderData, useMatches } from "@remix-run/react";
+import { render, screen, waitFor } from "@testing-library/react";
+import user from "@testing-library/user-event";
+import { createRemixStub } from "@remix-run/testing";
+import {
+  Form,
+  Outlet,
+  useActionData,
+  useFetcher,
+  useLoaderData,
+  useMatches,
+} from "@remix-run/react";
 import type { DataFunctionArgs } from "@remix-run/node";
 import { json } from "@remix-run/node";
 
 test("renders a route", () => {
-  let RemixStub = unstable_createRemixStub([
+  let RemixStub = createRemixStub([
     {
       path: "/",
       Component: () => <div>HOME</div>,
@@ -19,7 +27,7 @@ test("renders a route", () => {
 });
 
 test("renders a nested route", () => {
-  let RemixStub = unstable_createRemixStub([
+  let RemixStub = createRemixStub([
     {
       Component() {
         return (
@@ -45,16 +53,13 @@ test("renders a nested route", () => {
 });
 
 test("loaders work", async () => {
-  function App() {
-    let data = useLoaderData();
-    return <pre data-testid="data">Message: {data.message}</pre>;
-  }
-
-  let RemixStub = unstable_createRemixStub([
+  let RemixStub = createRemixStub([
     {
       path: "/",
-      index: true,
-      Component: App,
+      Component() {
+        let data = useLoaderData();
+        return <pre data-testid="data">Message: {data.message}</pre>;
+      },
       loader() {
         return json({ message: "hello" });
       },
@@ -63,9 +68,63 @@ test("loaders work", async () => {
 
   render(<RemixStub />);
 
-  expect(await screen.findByTestId("data")).toHaveTextContent(
-    /message: hello/i
-  );
+  await waitFor(() => screen.findByText("Message: hello"));
+});
+
+test("actions work", async () => {
+  let RemixStub = createRemixStub([
+    {
+      path: "/",
+      Component() {
+        let data = useActionData() as { message: string } | undefined;
+        return (
+          <Form method="post">
+            <button type="submit">Submit</button>
+            {data ? <pre>Message: {data.message}</pre> : null}
+          </Form>
+        );
+      },
+      action() {
+        return json({ message: "hello" });
+      },
+    },
+  ]);
+
+  render(<RemixStub />);
+
+  user.click(screen.getByText("Submit"));
+  await waitFor(() => screen.findByText("Message: hello"));
+});
+
+test("fetchers work", async () => {
+  let count = 0;
+  let RemixStub = createRemixStub([
+    {
+      path: "/",
+      Component() {
+        let fetcher = useFetcher<{ count: number }>();
+        return (
+          <button onClick={() => fetcher.load("/api")}>
+            {fetcher.state + " " + (fetcher.data?.count || 0)}
+          </button>
+        );
+      },
+    },
+    {
+      path: "/api",
+      loader() {
+        return json({ count: ++count });
+      },
+    },
+  ]);
+
+  render(<RemixStub />);
+
+  user.click(screen.getByText("idle 0"));
+  await waitFor(() => screen.findByText("idle 1"));
+
+  user.click(screen.getByText("idle 1"));
+  await waitFor(() => screen.findByText("idle 2"));
 });
 
 test("can pass a predefined loader", () => {
@@ -73,7 +132,7 @@ test("can pass a predefined loader", () => {
     return json({ hi: "there" });
   }
 
-  unstable_createRemixStub([
+  createRemixStub([
     {
       path: "/example",
       loader,
@@ -82,33 +141,29 @@ test("can pass a predefined loader", () => {
 });
 
 test("can pass context values", async () => {
-  function App() {
-    let data = useLoaderData();
-    return (
-      <div>
-        <pre data-testid="root">Context: {data.context}</pre>;
-        <Outlet />
-      </div>
-    );
-  }
-
-  function Hello() {
-    let data = useLoaderData();
-    return <pre data-testid="hello">Context: {data.context}</pre>;
-  }
-
-  let RemixStub = unstable_createRemixStub(
+  let RemixStub = createRemixStub(
     [
       {
         path: "/",
-        Component: App,
+        Component() {
+          let data = useLoaderData() as { context: string };
+          return (
+            <div>
+              <pre data-testid="root">Context: {data.context}</pre>
+              <Outlet />
+            </div>
+          );
+        },
         loader({ context }) {
           return json(context);
         },
         children: [
           {
             path: "hello",
-            Component: Hello,
+            Component() {
+              let data = useLoaderData() as { context: string };
+              return <pre data-testid="hello">Context: {data.context}</pre>;
+            },
             loader({ context }) {
               return json(context);
             },
@@ -130,18 +185,7 @@ test("can pass context values", async () => {
 });
 
 test("all routes have ids", () => {
-  function Home() {
-    let matches = useMatches();
-
-    return (
-      <div>
-        <h1>HOME</h1>
-        <pre data-testid="matches">{JSON.stringify(matches, null, 2)}</pre>
-      </div>
-    );
-  }
-
-  let RemixStub = unstable_createRemixStub([
+  let RemixStub = createRemixStub([
     {
       Component() {
         return (
@@ -154,7 +198,18 @@ test("all routes have ids", () => {
       children: [
         {
           path: "/",
-          Component: Home,
+          Component() {
+            let matches = useMatches();
+
+            return (
+              <div>
+                <h1>HOME</h1>
+                <pre data-testid="matches">
+                  {JSON.stringify(matches, null, 2)}
+                </pre>
+              </div>
+            );
+          },
         },
       ],
     },