+++ packages/remix-testing/index.ts
@@ -1,2 +1,2 @@
 export type { RemixStubProps } from "./create-remix-stub";
-export { createRemixStub as unstable_createRemixStub } from "./create-remix-stub";
+export { createRemixStub } from "./create-remix-stub";