+++ packages/remix-dev/CHANGELOG.md
@@ -1,5 +1,13 @@
 # `@remix-run/dev`
 
+## 2.1.0
+
+### Patch Changes
+
+- Sourcemap takes into account special chars in output file ([#7574](https://github.com/remix-run/remix/pull/7574))
+- Updated dependencies:
+  - `@remix-run/server-runtime@2.1.0`
+
 ## 2.0.1
 
 ### Patch Changes